Removed 3 unnecessary cast of "int"

+++ src/main/java/com/password4j/BcryptFunction.java
@@ -309,9 +309,9 @@ public class BcryptFunction extends AbstractHashingFunction
      */
     protected static byte char64(char c)
     {
-        if ((int) c >= INDEX_64.length)
+        if ( c >= INDEX_64.length)
             return -1;
-        return INDEX_64[(int) c];
+        return INDEX_64[c];
     }
 
     /**
@@ -407,7 +407,7 @@ public class BcryptFunction extends AbstractHashingFunction
         for (i = 0; i < 4; i++)
         {
             words[0] = (words[0] << 8) | (data[off] & 0xff);
-            words[1] = (words[1] << 8) | (int) data[off]; // sign extension bug
+            words[1] = (words[1] << 8) |  data[off]; // sign extension bug
             if (i > 0)
                 sign |= words[1] & 0x80;
             off = (off + 1) % data.length;
